desafio_sqlite_mongo_db/mongo/__init__.py
from os import environ

import pymongo
from dotenv import load_dotenv

load_dotenv()

client = pymongo.MongoClient(
    'mongodb+srv://pymongo:pymongo@cluster1.f9asews.mongodb.net/?retryWrites=true&w=majority'
)
