pyproject.toml
[tool.poetry]
name = "desafio-sqlite-mongo-db"
version = "0.1.0"
description = ""
authors = ["Carlos Eduardo <ducameneses16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.24"
pymongo = {extras = ["srv"], version = "^4.6.1"}
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.1.9"
isort = "^5.13.2"

[tool.ruff]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
